"feat: create getValueOfStudents() function"
